requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
fastapi==0.60.1
google-auth==1.20.0
h11==0.9.0
httptools==0.1.1
idna==2.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.6.1
python-dotenv==0.14.0
requests==2.24.0
rsa==4.6
six==1.15.0
starlette==0.13.6
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
